+++ src/main/java/com/breakoutms/lfs/server/undertaker/transfer/TransferRequestRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.breakoutms.lfs.server.undertaker.transfer.model.TransferRequest;
 
-public interface TransferRequestRepository extends JpaRepository<TransferRequest, Long> {
+public interface TransferRequestRepository extends JpaRepository<TransferRequest, Integer> {
 
 }
